Use function indices in MappedAOTBBlock.
Because HWT used symbol names to uniquely identify functions,
MappedAOTBBlock contained a &CStr to identify a function.

When SWT came along, it spoke function indices from the outset. We
shoe-horned it in to the existing infrastructure by converting the
function indices to symbol names.

This is unfortunate, because in order for the trace builder to look up a
function in the AOT IR it needs a function index. So SWT would start
with an index, convert it to a string, then convert it back to an index
(via a hashmap!). A pointless dance!

To avoid this, this change makes MappedAOTBBlock use a function index to
identify a function. In turn the nasty FUNC_NAMES hack (which allocates
a load of CStrings) can also go away.

Besides making the code simpler and more direct, it also speeds up the
trace compiler a bit. E.g. on Deltablue it consistently speeds up
"duration_compiling" by about 15%.