Better error messages and non-fatal gdb errors.
I've noticed that if you run some larger Lua programs under yklua with
debug assertions enabled, the gdb support code can crash due to maxing
out the file descriptor ulimit.

The reason for this is that, in debug mode, each trace writes textual IR
into a temp file so that gdb can show it in the source view. This makes
debugging JITted code a bit easier.

By default, Linux allows any given process to have 1024 fds, so it's not
too hard to max it out with gdb temp files.

This change:

 - Gives more specific error messages, so that you can see the
   underlying reason for the crash, and in the event of "too many open
   files" consider upping your `ulimit -n`.

 - Makes failure to write a gdb source file non-fatal. i.e. trace
   compilation is aborted, instead of the entire program crashing.
   An interested user can use YKD_LOG=3 to see why the trace was
   aborted.

Note that this is only ever an issue for debug builds of yk. gdb support
isn't built-in to release builds.