Don't run PT tests on machines that don't have PT.
Running this on an AMD Linux machine causes these tests to fail. There
isn't a great way to say "only run these tests on an Intel machine" so,
for now at least, unbreak things by having each test check for PT
support and returning immediately if it is not present.