Fix switch statements where multiple cases map to the same block.
This previously resulted in guards failing unnecessarily because we only emitted
a guard for the first case that matched the next block in the trace. Instead we
should match all cases that match the next block, similarly to how we deal with
the default case.

As a neat side-effect this allows us to simplify and deduplicate the switch
handling code a bit.