Replace two vague tests with a specific test.
This commit is about lang_tests that check "when we're outlinnig and we
promote something, the promotion data is ignored appropriately".

The two deleted tests both relied on trace_builder (or equivalent)
borking but that could happen for other reasons and the test would still
pass.

The new test makes sure the IR output is correct: it does that by first
calling an ignored promote, then a non-ignored promote. If the IR
contains the second value, then we've got a fairly strong test that
things are doing what we want.