Better handling of `yk_idempotent`.
Instead of waiting to see a call to `__yk_idempotent_promote_*()` and
retrospectively patching in the "idem const" into the corresponding
call, we can instead create the idem const at the time we encounter the
call.

After this there's no need for the `IdempotentPromote` opcode. The
ykllvm change means the AOT compiler no longer emits them.

This will help with Laurie's work on J2, which (by design) cannot
retrospectively patch past instructions.