Determine the trace optimisation level in MT.
Rather than having each JIT backend interpret `YKD_OPT` (which, perhaps,
should be `YK_OPT` but that's a decision for another day), record it at
the `MT` level. We could be very clever with what the levels mean
(should they be an `Option`? an `enum`? or ...?), but ultimately each
backend optimiser can interpret them differently: really only "0" and
"not 0" make sense at a global level. Hence defaulting to "1" is
probably good enough for now and quite a way into the future.