Allow the register allocator to be queried for output fills.
Consider this trace:

```
%0: ptr = reg
%1: i8 = load %0
%2: i64 = sext %1
```

Before this commit we would have to generate this code:

```
movzx x, [...]
movsx ..., x
```

i.e. we would zero extend a value from the heap, then immediately sign
extend it.

This commit introduces a variant of `RegAlloc::alloc` called
`RegAlloc::alloc_with_fills` which allows the caller to (a) tell the
register allocator that it can output certain types of `RegFill`s
directly (b) to be told which of those the register allocator wants it
to do.

The x64 backend can then optimise the code above to an immediate
load-and-sign extend single instruction such as `movsx x, [...]`.

In the future we might want to extend this to allow input fills to be
specified. Right now I don't have a concrete use case for that.