Standardise on the more flexible `build_and_test` function.
One thing that has always been a bit challenging for us in testing with
fm is that we have some necessary non-determinism in parts of the
system. For example, what register should the allocator choose to spill?
We can't sensibly specify all such details, but different choices can
then lead to different outcomes that are difficult to express in fm
tests. I tried extending fm, but the tests became too difficult to think
about: in essence, I'd reinvented the complexity of regular expressions
(but without confidence that the implementation would be correct!).

The solution to this I alighted upon a few days ago is obvious -- allow
tests to have multiple patterns that can be matched. That way we can
cover the nondeterminism but keep each test simple.

The register allocator was the first place I started with this, but I
wasn't sure if it was a good idea, so I allowed both "one pattern" and
"multiple pattern" tests. The latter is so easy -- it doesn't even
format on any additional lines -- that there's no good reason to keep
the former around.

This commit thus migrates the register allocator tests to all be
"multiple pattern" tests.