Use a RefCell<IsTracing> instead of a AtomicIsTracing.
Profiling benchmarks with the JIT disabled shows that the largest cost
is `__yk_trace_basicblock()` and a chunk of that is
`AtomicIsTracing::load()`.

This is because we currently use an atomic enum (inside a thread local)
to keep track of a thread's tracing state. The enum is already thread
local, so the atomicity is being used here solely for mutability.

Following a tip-off from perplexity, in an unexpected twist, a RefCell
around a regular enum appears to be faster than an atomic enum.

I tried three benchmarks with the JIT off for both configurations (note
the reduced benchmark parameters because the system is so slow with no
JIT):

```
Datum0: baseline
Datum1: refcell

confidence level: 95%

 Benchmark            Datum0 (ms)  Datum1 (ms)  Ratio  Summary
 richards/yklua/1       13865 ±83    10710 ±62   0.77  22.76% faster
 json/yklua/1            2708 ± 8     2093 ± 7   0.77  22.71% faster
 deltablue/yklua/200     2604 ± 3     2014 ± 7   0.77  22.67% faster
```

I tend to believe the summaries, since the CIs are so narrow.

I then ran the whole suite with the JIT on to check for regressions:

```
Datum2: jiton-baseline
Datum3: jiton-refcell

confidence level: 95%

 Benchmark                 Datum2 (ms)  Datum3 (ms)  Ratio  Summary
 LuLPeg/yklua/             14619 ± 206  14273 ±  42   0.98  2.37% faster
 towers/yklua/600           8832 ±  60   8690 ±  48   0.98  1.61% faster
 fasta/yklua/500000         3665 ±  16   3620 ±  11   0.99  1.25% faster
 fannkuchredux/yklua/10    12216 ± 172  12426 ±   5   1.02  1.72% slower
 richards/yklua/100        41152 ± 432  41995 ± 293   1.02  2.05% slower
 HashIds/yklua/6000        10402 ±1011   9859 ±  49   0.95  indistinguishable
 queens/yklua/1000          4919 ± 250   4688 ±  58   0.95  indistinguishable
 sieve/yklua/3000           3862 ± 197   3721 ±  17   0.96  indistinguishable
 nbody/yklua/250000         5113 ± 250   4962 ± 121   0.97  indistinguishable
 permute/yklua/1000         8116 ± 509   7926 ± 237   0.98  indistinguishable
 cd/yklua/250              33757 ± 716  33279 ± 620   0.99  indistinguishable
 json/yklua/100            13057 ± 368  12928 ± 161   0.99  indistinguishable
 list/yklua/1500            7912 ± 577   7864 ± 363   0.99  indistinguishable
 havlak/yklua/1500         74775 ± 657  74473 ± 273   1.00  indistinguishable
 knucleotide/yklua/        14163 ± 191  14133 ± 279   1.00  indistinguishable
 mandelbrot/yklua/500       1272 ±   2   1271 ±   1   1.00  indistinguishable
 storage/yklua/1000        23217 ±  65  23206 ± 254   1.00  indistinguishable
 spectralnorm/yklua/1000    8953 ±   1   8960 ±   9   1.00  indistinguishable
 binarytrees/yklua/15      11315 ±  66  11349 ± 110   1.00  indistinguishable
 deltablue/yklua/12000     10607 ±  25  10664 ± 135   1.01  indistinguishable
 Heightmap/yklua/2000       7133 ±  21   7250 ± 187   1.02  indistinguishable
 bounce/yklua/1500          8784 ±  86   8982 ± 291   1.02  indistinguishable
 bigloop/yklua/1000000000  23602 ± 151  24224 ± 510   1.03  indistinguishable
 revcomp/yklua/             3415 ± 534   3760 ± 232   1.10  indistinguishable
```

This was with 3 process executions.

I then ran the benchmark with the most "slowdown" with more process
executions (10) to see if that effect is real:

```
Datum5: jiton-baseline-richards
Datum4: jiton-refcell-richards

confidence level: 95%

 Benchmark           Datum5 (ms)  Datum4 (ms)  Ratio  Summary
 richards/yklua/100   41391 ±447   41252 ±276   1.00  indistinguishable
```

So I think this change improves interpreter, but not overall,
performance. At least for these benchmarks.