Flatten the two "put const in register" functions.
It turns out that the function are misleadingly named: `zero_fill_const`
was also used for `RegFill::Undefined`. Rather than make a third
function, it makes more sense to have the backend choose based on a
`RegFill`. This is also in-line with other similar functions the
backend has to implement.