Rethink how/when constants are put in registers.
Previously the `post_inst` function was a hack to put constants in
registers at a point that made sense in jit-asm output from a human
perspective. However, it did so eagerly: so if `%2` and `%1` both
referenced a const in `%0`, one could get multiple "put the constant in
`%0` registers", even when they weren't needed. Besides that,
`post_inst` just felt... ugly.

This commit rethinks that. It makes register allocation for
[Inst::Const]s an explicit, rather than an implict, thing. This reduces
a fair big of the ugliness -- though we have to introduce a new, but at
least fast, function `alloc_const` -- and means that constants are now
only put into registers when they really need to. In many cases, that
means that only when the `Inst::Const` is actually encountered does this
happen, which is what we want (and checked by the
`constants_as_late_as_possible` test).

It also has the benefit that "put a constant in a register" now appears
once (in `asm_ractions`) and not twice, as before. Given some changes
we're going to want to make soon, this avoids code duplication.

While doing this, I realised that our specification of "merge
instruction values" was both vague and unnecessarily fussy. The doc
string at the top of the module reflects what I now believe it should
be, but it does not simplify across the module as should be possible:
that's for a future commit.