Add runtime support for callee shims.
According to benchmarking, using shims speeds up the interpreter by
about 30% and even some benchmarks with the JIT enabled too.

Crucially for this yk-side change, if the trace builder encounters a
call to a shim, it should "look through" the shim (which is marked
yk_outline) and inline the unoptimised version of the function into the
trace.