Add support for float and double consts.
This touches a lot of stuff: from the HIR parser onwards. Not all of
this commit is quite as profound as it looks: in particular, the changes
to `i_load` and `i_store` really involve splitting each up into
int/float variants.

In a sense, the core change (beyond the necessary boilerplate) is to
allow x64 immediates to go to a general purpose register and then
bitcast to an XMM register. That's very simple because of the preceding
commits. But in order to test all the interesting use cases this throws
up, we have to add a lot of extra stuff, hence the overall size of the
commit.