Add support for calling functions with floating point args / returns.
There's quite a bit of boilerplate here (e.g. support for spilling /
unspilling floating point numbers), but the most interesting parts are
in `cg_call` in the x64 backend. My first attempt at this made this
fiddly bit of code even worse, because we had combinatorial explosion of
sorts when floating point arguments and var args combined.

But I then took a step back, and -- although it's now of course longer,
because it has to deal with floating point numbers -- it's now much
simpler to understand. I could probably simplify it even further (and
there's at least one missing optimisation), but it's more than good
enough for now. The tests for this are somewhat voluminous, because
there are a lot of combinations to consider in the SysV x64 ABI.