Force a reload of callee saved registers.
This reflects the changes back in
bc59d8bff411931440459fa3377a137e8537a32f: callee saved registers will
always also be spilt, and coming into a trace they are potentially
corrupted, and must be reloaded.

The fix in this commit is simple, but ugly, and won't work with some
future expansions of j2 that we know we want to do. However, doing a
proper job turns out to be quite a bit of work, and would block what we
care most about right now.