For signed operations, operate on 64-bit registers.
`RegFill::Signed` means "any unused bits are sign extended" but if we do
a signed operation on a 32-bit register, the upper 32-bits will be
zero-extended. That means that `sub r.32._, r.32._` requires us to set
`out_fill: RegCnstrFill::Undefined` but we actually set things to
`RegCnstrFill::Signed`.

There are cases where we'd like `Signed` and others `Undefined`. j2's
register allocator does give us the tools to do that, but at the moment
we're in the worst of all worlds: we' do the wrong thing in some cases.
The quickest, and easiest to understand, fix is to always operate on
64-bit registers which is what this commit does. With that
`simple_binop` passes (modulo the somewhat pointless jitc_yk specific
`jit-pre-opt` part of the test).