A better heuristic for the size of the assembled buffer.
Before this commit we requested a fixed sized buffer from the OS for the
assembled buffer: if the assembled trace exceeded that, hilarity
occurred.

This commit implements a simple heuristic that at least takes the actual
trace into consideration. In essence, we guess a number of bytes needed
per HIR instruction, over-allocate (by ~50%), and after assembly return
any wasted space to the OS.

[In case it's not obvious, we have no choice but to preallocate a buffer
on x64 if we want to optimise near calls: we have to know the address of
the buffer to work out if a call can be assembled to a near call or
not.]