Only optimise `eq`.
This is what happens when I quickly hack an optimiser: I make silly
mistakes! Still, these are easily fixed. In essence, we can only learn
that x==y if there really is an `==` (and not, say, `!=`).

A proper, tested, optimiser is on my todo list, but there are higher
priority items right now, so this quick fix will have to do.