Introduce a `Cast` register constraint.
In LLVM IR, casting can often be a no-op: for example, `trunc` can
always be a no-op; and `zext` can sometimes be a no-op. In order for
this to happen, our register allocator needs to be able to merge the
values of multiple instructions into a single register.

We've had support for this from the beginning, but it was unused (except
for entry variables) because we previously modelled casting as an
`[InputOutput]` register constraint. In code such as:

```
%0: i8 = arg ...
%1: i16 = zext %0
blackbox %1
blackbox %0
```

this would cause a `Copy` (or spill) such that `%0` and `%1` lived in
two registers.

This commit introduces `Cast` constraints: they allow code in the
example above to avoid a copy/spill i.e. `%0` and `%1` live in the same
register.

The practical effect of this is somewhat limited right now by the fact
that the register allocator is spill-happy: when it encounters a value
in the "wrong" register, it spills it, rather than moves it. When we fix
that, we should be able to use the presence of `Cast` instructions to
identify merges, helping us allocate registers more effectively. Still,
it feels that (hopefully) getting this API right now is the correct
sequencing.