Implement more of `VarLoc::StackOff`.
Although it's not obvious at first, `VarLoc::StackOff` is, in practise,
somewhat close to a constant: at any point we can easily derive it from
the frame register (RBP on x64). It's a good candidate for spilling
(we don't need to actually spill it!) and unspilling (we don't need to
do a load, just a simple add/sub relative to the frame register).

This is enough for the `ptrtoint` test to run.