Remove compile-time guards.
These are always, and will almost certainly always be, on -- until we
remove a backend at which point we'll remove all the relevant code.

At the moment their only practical effect is to confuse rust-analyzer
(causing it to sometimes stop completing things in one other subsystem),
which isn't a price that seems worth paying.