Implement `uitofp`.
This turned out to be surprisingly fun! In essence, the fastest way of
doing this involves some magic constants, so we need a way to record
constant data in a compiled trace. This commit therefore introduces an
x64 "data section", which we model as a block, knowing that starts with
16 byte alignment.