Trace previous `BBlockId`s properly.
Before this commit we (briefly, but importantly) lost context of the
previous `bid` when entering outlining, which meant that we didn't catch
all of the irregular control flow we should.

I'm not particularly happy with the design of this commit: it's fragile
and easily misunderstood. However, I haven't immediately been able to
come up with something superior, so this feels better than nothing.