Hack in a fix required for a block-splitting optimisation.
`prev_bid` doesn't take into account that we could have just returned
from a function, meaning that it's possible to have a `prev_bid` from
another frame (and therefore function), but the intention is that
`prev_bid` is the previous BBlockId for the *current* frame.

This is a horrible hack. We expect trace_builder to be removed in favour
of j2 in the nearish future, so we will tolerate it for now.

Initial benchmarking suggests a few % speedup for a couple of
benchmarks.