Use a two-stage `dlsym` cache.
Caching `dlsym` is somewhat challenging, because it can return different
results depending on what thread it's being run on (for "thread
locals").

This commit addresses this in three ways:

  1. It puts a non-thread-local cache at the j2 top-level.

  2. It forces users to specify whether `symbol` is thread-local or not.
     j2 always knows this information.

  3. The cache is two-stage. Each thread maintains its own cache, but
     there is a global cross-thread cache. Thread-locals stay in the
     per-thread cache. Non-thread-locals can be copied from the global
     cache.

That third point has another advantage: it means that most `dlsym`
lookups don't need to grab a mutex for the global cache. If someone runs
j2 on a machine with a monster number of threads, that means that
performance should not be blocked on the global cache.