Try to `mmap` memory near the interpreter and existing traces.
If we leave `mmap` to its own devices, it will often (but, at least on
Linux, not always!) put traces very
far in virtual memory from the interpreter. That means we end up having
to do inefficient far calls from traces back to the interpreter.

This commit takes inspiration from LuaJIT: we try to force `mmap` to
allocate traces near to the interpreter. We use `mmap`'s "hint" feature:
if you give it an address, it will "try" to allocate memory nearby in
virtual memory. However, if nothing nearby is available, it will give
up, and allocate somewhere completely different. What we do is to look
at the address we've got back and see if, on our current platform, the
address is nearby; if not, we `munmap` the memory, and try a different
hint address.

This works quite well for yklua. My guess is that for interpreters with
a much greater quantity of code (e.g. CPython / CRuby), we might have to
fiddle with the "try another hint address" heuristic, but that's a
bridge we can cross when we get to it.