Split the optimiser into parts that are plausibly maintainable.
Before this commit, j2's current optimiser was a minimal test that
something like this would work: it was comically incomplete, untested,
and internally unstructured.

This commit starts to get j2's optimiser into some sort of shape. It
breaks strength reduction and constant folding into its own module, and
makes it testable. I've tried to avoid one of jitc_yk's notable
overheads by removing the concept of "decopy_eq" (we could sometimes
spend 40% of compilation time in such functions!) by using `map_iidxs`
and a new `OptOutcome` enum. I haven't fully convinced myself that this
will handle everything we need in the future, but until I've got further
into this, it's hard to know for sure --- and it is simple!

So as not to overload this commit, this can currently only optimise
`and`: we can add further optimisations in subsequent commits.