Fix the topological sort nodes / edges.
Long story short: I got what should be nodes and edges in the
topological sort the wrong way around and this commit fixes them,
and adds a test that identified the mistake.

Slightly expanding: registers are now nodes and `RegCopy`s define edges.
The core topological sort algorithm was correct before, but now it's
operating on the right things! There is one additional subtlety: we want
the register _copies_ to be in order, not the _registers themselves.
When the topological sort succeeds, it produces sorted nodes, which we
can then use to sort the edges.

Note: because there are often multiple valid topological sort, this can
lead to the order that registers are copied in changing (and you can see
that in a couple of tests below).