Implement guard optimisation.
This showed up something I should have considered in the API earlier:
void instructions can't be transformed into another instruction, they
can either be inserted as-is or not inserted. So part of this commit is
splitting `feed` and `feed_void` apart; the other part is the actual
guard optimisation.

I'm still not happy with how we represent copying at the moment: we miss
optimisation opportunities, and the API just doesn't feel right. But
fixing that is arguably a second step that can be done in a future PR.