Use a better heuristic for the initial `mmap` hint address.
Even on the fairly small yklua, we previously had to retry a number of
times with `mmap` on our first attempt to find a sensible hint address.
For larger VMs, this will be even more painful.

There's no platform-independent way of working out what a good hint
address is, but this commit uses a simple proxy: `main +
sizeof(executable)`. This is fairly quick to calculate, and gives us a
plausible hint address straight away. On yklua on Linux this reduces the
retries down from 8 to 0.