Remove guard optimism.
There is a subtle bug in the way this is currently implemented and
fixing it would be a pain. As I intend a more generic optimisation which
will subsume this, it doesn't seem worth carrying around all this code
(even if it's fixed). This does slow down a small number of benchmarks
by 2-6%, so it has an impact, but small.