Add an explicit `int_min_poison` to the HIR parser.
The HIR parser now accepts both:

```
%0: arg [reg]
%1: abs %0
%2: abs %0, int_min_poison
```

where %1's intruction sets `int_min_poison` to false, and %2's to true.

That requires tweaking a few tests for cosmetic purposes.

This highlighted to me that for C, `int_min_poison` is true, and that
our x64 backend may not be correct if it is false. I have thus
explicitly added an assert to that effect: if we encounter false at some
future point, we can rethink the code generation.