Move the top-level optimiser to only knowing about instruction equivalence.
I always intended `Range` here to be a placeholder for thinking about
this a bit more: I have now thought about it a bit more, and I think
"range" is the wrong thing for the top-level optimiser to know about.
All we really need to know is "at this point during optimisation, do we
know that `InstIdx` x can now be considered equivalent to y?"

This commit thus gets rid of `Range` as a top-level thing and defines a
simpler notion of "equivalence".