Allow equivalence of non-consts.
If we have `icmp %7, %8` followed by a guard, then we can arbitrarily
pick `%7` or `%8` as a "winner" henceforth: this stops us carrying
around one more live value.

This means that it makes the code somewhat clearer if we move the
"set equivalency" check into `set_equiv`. This means that, because we
currently tend to duplicate constants, we can end up with seeming
inequivalence. This is the first good argument for deduplicating
constants that I've recognised.