Make canonicalisation and rewriting of iidxs take `&mut self`.
Previously we separated "rewrite iidxs" from "canonicalisation" but in
the main optimiser, we always do these together: at least currently,
it's only in testing that we need a separate notion of "rewrite iidxs".

In an ideal world, we'd probably make `canonicalise` take a generic
function, but that doesn't work because we want to call it with
`equiv_iidx` which takes `&mut self`.

This commit thus makes `canonicalise` do all operand rewriting: in
normal code one thus only needs to call `canonicalise`.

We then provide a subset-function `rewrite_iidxs` which does take a
generic function and which is only intended for testing purposes.

This inevitably duplicates some code, but it's currently the least worst
design I can think of.