Avoid copying instructions all the time.
The initial aim of this commit is to avoid instructions like `rewrite`
which consume and produce `Inst`s (i.e. copying them, or at least
placing them on the stack temporarily): instead we prefer to call
`canonicalise` which mutates instructions in the heap.

That has more of a knock-on effect than I initially hoped for a single
commit. It means that I needed to introduce `as_constkind` which allows
the optimiser to make getting at constants (if they're referenced!) much
easier.

The end effect of this commit is that we have slightly faster, and
noticeably smaller, code.