Guarantee that [Ty]s are not duplicated.
Previously, we allowed the same [Ty] to be inserted multiple times. This
is obviously inefficient, but it also means that type equality requires
a full `PartialEq` of the struct.

This commit changes all of the HIR creators -- and adds a check in
`assert_well_formed` to check this has been done -- that a given [Ty] is
inserted only once. This allows type comparison to be integer
comparisons.