Implement common subexpression elimination.
This commit is a mixed bag: the implementation of CSE is, mostly, very
efficient, thanks to a couple of tricks (one of which definitely comes
from LuaJIT); but there are some horrors left behind (the FIXMEs).
However, leaving this even longer doesn't seem the right trade-off:
this commit is already bigger than is comfortable to review.

Mostly that is due to the `cse_eq` functions each HIR instruction needs.
The `cse.rs` module itself is fairly small.