Rework the internal optimiser API.
The previous API was a rush job, and not a good one at that: it had no
explicit notions of "passes", and had to treat `cse` differently than
`strength_fold`; and it presented the same API to `aot_to_hir` as to
"passes".

The main thing this commit does is to split the optimiser such that it
has two different APIs: an outward-facing API for `aot_to_hir` and an
internal-facing API (`PassOpt`) for optimisation passes. Both
`strength_fold` and `cse` are now "proper" passes (implementing a trait
`PassT`) on an equal footing.

The way this is achieved is by having the outward and internal facing
objects share a `OptInternal` struct. The main advantage of this is that
it allows us to assuage the borrow-checker. However, it also indirectly
means that we can expose different functionality in the outward and
inward-facing APIs. That does mean some changes in `hir`, in particular
the introduction of a new trait `EquivIIdxT`, which allows us to pass
much "smaller" objects around (i.e. not full optimisers!).

That allows this commit to deal with one of the true horrors of the
previous situation: `strength_fold` could insert instructions into the
trace. These are now "preinstructions" and the API makes it much harder
for them to be used incorrectly (as well as documenting why/where one
should think carefully about them).

Overall I think this commit shoves us a long way in the right direction.
There is still work to do though -- notably `OptOutcome` as it currently
stands doesn't allow a pass to hand-over knowledge it has about an
instruction. However, this commit is big enough as-is, and I haven't yet
fully worked out what the next step would be. So it feels better to
consider this now, rather than wait ever longer for perfection.