Extend the `inst_commited` API.
For load/store elimination, we need to know the `iidx` of the to-be
committed instruction, and access to the optimiser for `equiv_iidx`ing
and so on. This commit thus makes `inst_commited`:

```rust
fn inst_committed(&mut self, ci: &CommitInstOpt, iidx: InstIdx, inst: &Inst);
```

where `CommitInstOpt` is vaguely similar to `PassOpt` _except_
`CommitInstOpt` is `&` (not `&mut`) and therefore doesn't allow types /
instructions to be pushed. I would like to pretend that I anticipated
this API would naturally fall out of the hat from cf51a9e3 but I did
not. Happily, it does fall out of the hat!