Don't generate pointless self copies.
`self_copies` served two purposes:

1. Telling us when register fills have changed.
2. recording that a register is used.

However, we don't make use of (2) yet, there are lots of cases where a
register is used and unchanged, and `RegCopy` is a heavy-weight struct.
All that means that we are putting a lot of values into `self_copies`
without deriving much benefit from doing so.

This commit recognises that we don't need to generate "fill changes" if
fills haven't changed or if the destination fill is `Undefined`. Doing
so saves about 5% of compilation of time, probably mostly because in
about 90% of cases we no longer `malloc` anything.