LLVM's `memcpy` has slightly different outcomes compared to libc.
libc's `memcpy` gives `memcpy(x, x, ...)` undefined behaviour, but
LLVM's `memcpy` gives this defined behaviour. The outcome from our
perspective is the same in either case (we can remove the `memcpy`), but
the reason is different.

While here, we can generalise the "equivalence" check to whenever the
SSA variables are the same, even if they're not constant.